improve readability, refactor MGSwipeCard and State classes
