--- Example/MGSwipeCards/SampleCardModel.swift
@@ -1,16 +1,0 @@
-//
-//  SampleMGSwipeCardModel.swift
-//  MGSwipeCards
-//
-//  Created by Mac Gallagher on 6/2/18.
-//  Copyright © 2018 Mac Gallagher. All rights reserved.
-//
-
-import UIKit
-
-struct SampleCardModel {
-    let name: String
-    let age: Int
-    let occupation: String?
-    let image: UIImage?
-}
