+++ Example/MGSwipeCards/UIView+Anchor.swift
@@ -1,5 +1,5 @@
 //
-//  Utilities.swift
+//  UIView+Anchor.swift
 //  MGSwipeCards
 //
 //  Created by Mac Gallagher on 6/2/18.
@@ -33,15 +33,4 @@ extension UIView {
             heightAnchor.constraint(equalToConstant: height).isActive = true
         }
     }
-    
-}
-
-extension UIColor {
-    
-    static var sampleRed = UIColor(red: 252/255, green: 70/255, blue: 93/255, alpha: 1)
-    
-    static var sampleGreen = UIColor(red: 49/255, green: 193/255, blue: 109/255, alpha: 1)
-    
-    static var sampleBlue = UIColor(red: 52/255, green: 154/255, blue: 254/255, alpha: 1)
-    
 }
