+++ Extensions.swift
@@ -1,12 +1,38 @@
 //
-//  UIView+Anchor.swift
+//  Extensions.swift
 //  MGSwipeCards
 //
-//  Created by Mac Gallagher on 6/30/18.
-//  Copyright © 2018 Mac Gallagher. All rights reserved.
+//  Created by Mac Gallagher on 10/8/18.
 //
 
 import UIKit
+
+extension CGPoint {
+    
+    func distance(to point: CGPoint) -> CGFloat {
+        return sqrt(pow(point.x - self.x, 2) + pow(point.y - self.y, 2))
+    }
+    
+    // Returns the point in the coordinate system -1 <= x,y <= 1 with the same relative position as in the provided frame
+    func normalizedDistance(forSize size: CGSize) -> CGPoint {
+        let x = 2 * (self.x / size.width)
+        let y = 2 * (self.y / size.height)
+        return CGPoint(x: x, y: y)
+    }
+    
+    func dotProduct(with point: CGPoint) -> CGFloat {
+        return (self.x * point.x) + (self.y * point.y)
+    }
+}
+
+extension Array {
+    
+    func shift(withDistance distance: Int = 1) -> Array<Element> {
+        let offsetIndex = distance >= 0 ? self.index(startIndex, offsetBy: distance, limitedBy: endIndex) : self.index(endIndex, offsetBy: distance, limitedBy: startIndex)
+        guard let index = offsetIndex else { return self }
+        return Array(self[index ..< endIndex] + self[startIndex ..< index])
+    }
+}
 
 extension UIView {
     
@@ -43,5 +69,4 @@ extension UIView {
         
         return anchors
     }
-    
 }
