+++ MGSwipeCards/Classes/MGCardStackView/MGCardStackViewDataSource.swift
@@ -6,10 +6,7 @@
 //  Copyright © 2018 Mac Gallagher. All rights reserved.
 //
 
-import Foundation
-
 public protocol MGCardStackViewDataSource {
-    
     func numberOfCards(in cardStack: MGCardStackView) -> Int
     func cardStack(_ cardStack: MGCardStackView, cardForIndexAt index: Int) -> MGSwipeCard
 }
